backend/requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiogram_dialog==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.9.2
attrs==25.3.0
billiard==4.2.2
cachetools==5.5.2
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
dotenv==0.9.9
drf-yasg==1.21.11
frozenlist==1.7.0
idna==3.10
inflection==0.5.1
isort==6.0.1
Jinja2==3.1.6
kombu==5.5.4
magic-filter==1.0.12
MarkupSafe==3.0.3
multidict==6.6.4
packaging==25.0
prompt_toolkit==3.0.52
propcache==0.3.2
psycopg2-binary==2.9.10
pydantic==2.11.9
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
yarl==1.20.1
